Staging
Did the staging, like I said I would. Quit blackmailing me, unless you
want me to hold more of a grudge against you. The way I made the staging
was like it's the main function. I also made it to where it would be
better to use the old rocket calculations, because i just made alpha=0
for when the stages break off.