Fixed Staging
Had it to where it was updating y instead of z in position for {x,y,z}.
Forgot that z was the height and y was depth. Also just to visualize,
made theta in manipulate for all animations go from 0 to pi/2 because we
don't really need the negative values. Working on a function that has a
vertical asymptote at x=0 and y=*desired low earth orbit* to manipulate
theta at least as a constant function. The PowerPoint will also be done
by 8:00.